+++ src/index.ts
@@ -5,6 +5,7 @@ export {
     any,
     contains,
     enumerate,
+    every,
     filter,
     iter,
     map,
@@ -12,6 +13,7 @@ export {
     min,
     range,
     reduce,
+    some,
     sorted,
     sum,
     zip,
@@ -50,4 +52,4 @@ export {
     uniqueEverseen,
     uniqueJustseen,
 } from './more-itertools';
-export { compact, compactObject, first, flatmap, icompact } from './custom';
+export { compact, compactObject, find, first, flatmap, icompact } from './custom';